Move DefaultSystemMessage to a separate file
